Ignore unindexed apps in app score calculation
Previously apps that had indexed instances, but were not
themselves indexed would be included in the app score calculation.

This amounted to about a 50% discrepancy in the app score total
we expected to see.

Unindexed apps are mostly core apps such as APM registries,
APM repositories, APM ENS Sub registrars and the like.