Remove worker and taskmaster
The worker was the actual indexing component of Apiary,
and the taskmaster was the component that queued units of work
for the workers to process in order to facilitate multiple
workers running at once in parallel.

These components are being replaced by a subgraph and the API
will be a wrapper around one or more subgraphs instead.

The documentation has been updated accordingly.