Add basic API based on The Graph
This version of the API decorates aspects of the Aragon Connect
subgraph, however it currently lacks some information, most
notably:

- The address of the kit used to create the organization,
  which is unavailable in the subgraph we are currently using.
- The ENS name of the organization, which is unavailable in the
  subgraph.
- A creation timestamp, which is unavailable in the subgraph.
- An AUM metric which we will need to get from another
  subgraph.
- The `updateProfile` mutation currently present in Apiary,
  however this is somewhat trivial to add.

However, it does contain the ability to read existing profiles
in our database.

This is a major breaking change, since most of the schema changed,
including the pagination and filtering mechanism, as well as
a lot of the field names and types.