Make profiles wiki-style by default
If a profile is not claimed by one or more editors then anyone
can contribute to the profile of an organisation, provided
they pass a valid signed description of the changes they wish to
make.

A profile is considered claimed once an address with permissions
on the specific organization makes a change to the profile.